Added coroutine, cache, limit update entity and render (#196)
* Added coroutine, cache, limit update entity and render

improved healthbar performance, 
This commit part of main my private fork special for pull request.

* Refactoring

* Deleted ValueTuple

* Refactoring waits for coroutine

* Fixed miss interface after refactoring

* Deleted GameIsLoading

* added offset pattern of isLoadingScreen

* RangeNode Event OnValueChanged

* Replaced F# Tools to C#

* Updated coroutine and gamecontroller for that

Added one more runner in parallel thread
Reworked update coroutine settings

* Fix entitylabel

* Improved speed start hud

* Rework main cache

* Little performance, replaced tons of lists

* Little improved performance for heatlhbar

* Cache components for entitywrapper

* XpRate Areachange wait while loading

* Dont initialise() plugins if it load disabled

* Added plugins settings information for coroutine logic

Resume/Pause when plugins Enable/Disable

* Little fix for useless loop

* Fix InvalidOperationException when pickup items very fast

* Change path to address for id, better for performance

* Load performance settings before plugins for can be their use it

* Added settings in menu

Didn't test

+++ src/Program.cs
@@ -9,7 +9,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Windows.Forms;
-using Tools;
+using PoEHUD.Tools;
 
 namespace PoeHUD
 {
@@ -117,7 +117,7 @@ namespace PoeHUD
                 sb.Append(System.Environment.NewLine);
 
 
-                sb.Append("Area change: " + memory.ReadLong(offs.AreaChangeCount + memory.AddressOfProcess));
+                sb.Append("Area change: " + memory.ReadInt(offs.AreaChangeCount + memory.AddressOfProcess));
                 sb.Append(System.Environment.NewLine);
                 sb.Append(System.Environment.NewLine);
